all: many fixes
Many fixes and tweaks based on efforts to build
a bootstrap toolchain. One notable change in particular
is that plan-inputs is now an alist where the car of
each pair indicates the directory that acts as the root
in which the cdr items are extracted. (Thus, 'inputs' are
extracted to /sysroot while 'tools' are extracted to /, and so on.)