execline: automatically escape strings
Fixes two related issues:

First, in order to guarantee that a single scheme string
became a single execline word, you had to manually put
quotes around it. Just do this automatically when the
string would need quotes in order to be one "word."

Secondly, complain loudly when being asked to print
complex numbers. The expression '(sed -i) is a symbol
and a complex number, not two symbols.