sysplan: even more import trickery
Avoid the (eval '(import-for-syntax ...)) import library bits.
We're just interesed in the module registration code.
(Technically we'd need to inject the import-for-syntax bits
when the modules are imported into the lexical environment,
but we only ever define macros that depend on the simplest
top-level scheme environment.)

Also fix a repl history bug in the sysplan repl.