sysplan: fix/improve loading of (pkg ...) forms
Chicken refuses to (load ...) module forms from within
other module forms (it complains that the module is not being
evaluated in the top-level environment), so we have to perform
our own recursive pre-loading of required modules. This guarantees
that by the time chicken sees (import (pkg perl)), the (pkg perl)
module is already loaded, and so it only has to import bindings.

This ends up being a lot faster than a fancy er-macro-transformer anyhow.