pkgs: make modules implicit
A package of the form (pkg foo) is implicitly wrapped in

(module (pkg foo) (foo) <body> ...)

This simplifies the loading code and cuts down on some boilerplate.
It also implicitly enforces that (pkg foo) export a symbol 'foo,'
so now there's an obvious way to go from package name to package-lambda.

The ssl-client package had to be split from libressl.scm as a consequence.