plan: only deal in artifacts
There's no reason that (plan) needs to know anything about recipes.
The /env and /build files are just files, so they can just
be represented as artifacts. The recipe logic is moved into
(package) where it belongs.

The implicit contract between (plan) and the rest of the world
is merely that /build is an executable, /env is a file that can
be loaded by /bin/envfile, and some bits of execline-tools
(redirfd, fdmove, export) are in $PATH after /env is loaded.

Another upshot of this approach is that loading old plans
is even simpler; they *only* contain references to artifacts.
This ought to ease forwards-compatibility of old plan hashes.

Tested with scripts/rebootstrap and scripts/build-world.