plan: get rid of 'fetch-remote-file!' hack
There wasn't (previously) a way to specify remote files
like remote archives, since the 'extra' artifact field
was used for inline data. Now that field has a pair in it
which indicates whether or not there is an inline string
or a source url.

Like remote archives, fall back to the CDN when no url is available.