scripts/reproduce: new script
The 'reproduce' script takes an existing plan hash,
loads it into a plan, and then executes it
(and compares it to known results if we have them).
The plan has to be in the plan database already.
It does not need to have a known output, and it
may not need local copies of its dependencies,
provided they can be pulled from the fallback CDN.

Thus you could do a (slow) consistency check
on an entire plan database without access to the
code that generated the plans.

Also just useful for spot-checking plan repros.