image: new module
Add 'initramfs' and 'squashfs' image-producing functions
that take a list of input plans and combine them into
a single output image.

Had to do a bit of hacking in plan.scm to make this work well,
because we don't want the initramfs or squashfs image to get
stuck in a (poorly-compressed) tarball. So plan#plan now has
a 'raw-output' option that specifies a single file in /out
is the actual plan output.

A bunch of packages in pkg/ got moved into image.scm because
they were dependencies of squashfs-tools that weren't
already in base.