all: simplify graph traversal transformers
Introduce a 'kompose' macro that lets the user
bind arbitrary labels to specific points in the
reducer so that it is straightforward to define
recursive reducing functions.

Also, export 'busybox/config' from base, which
will create a busybox package from a config
so that we can use a more full-featured busybox
install in the (forthcoming) init-system code.

Also, update linux-* 5.4.15 -> 5.4.18