service: check in some code
Initial check-in of the service code.
(It is untested and probably wildly broken,
but this change at least keeps *existing* stuff
working.)

A "service" sits one level of abstraction above
a "package," because a service resolves to
mulitple packages. More specifically, a list
of services resolves to a list of packages
including the packages representing an init system
that starts the specified services.

The workflow for building a bootable system image
would be something like

  (mksquashfs (services->packages services ...))

This change also pulls s6, s6-rc, and linux into
builtin packages, and introduces a larger version
of busybox that has the utilities a bootable linux
system would require.