linux: upgrade to 5.4.20
Also, make plan#build-graph! recognize plans with
parallel marked 'very and assign them as many
resources as possible. (This is for packages like
a linux kernel or gcc that take long enough to build
that it's heuristically worth waiting for other builds
to finish in order to use more resources. Presently,
the heuristic is just "wait for all resources to become
available and then take all of them," which is good
enough for now.)