plan, coroutine: implement a GNU make(1) job server
When GNU make invokes sub-makes, it uses a pipe
as a semaphore to throttle jobs (from -j <jobs>).
Those file descriptors are passed via MAKEFLAGS
as "--jobserver-auth=r,w".

This patch removes all of the parallelism heuristics
from plan.scm, package.scm, etc. and replaces it
with a jobserver implementation. All make(1) invocations
inside the build sandbox share the same jobserver pipe fds,
so we no longer have to use the $nproc hack to get intra-build parallelism.

GNU make expects the pipe to have O_NONBLOCK set, so we have
to handle EAGAIN when reading from the pipe. Unfortunately
this meant hand-rolling a bunch of fd juggling in coroutine.scm
to handle scheduling around non-blocking file descriptors.
The upside of this is we no longer need the semaphore code.
The pid-reaping code gets a bit of internal refactoring to
use the self-pipe trick on SIGCHLD.

(I put off making this change for a while, because I wasn't
sure that enough packages would use make(1) to be worth it.
Turns out basically *everything* uses make(1), so it's totally worth it.)