linux, fs, base: work for supporting platforms that use u-boot
- Fix linux/config when device tree blobs need to be built.
The set of fixes for dtc are nearly identical for this and u-boot.
- Port u-boot; expose uboot/config. This needs some moderate patching
around dtc, similar to linux/config.
- Create a kind of hacky patch for mke2fs so that we can generate
deterministic ext2 filesystem images.

I've also ported ATF ("Arm Trusted Firmware") for the macchiatobin,
but it needed so many target-specific hacks in order to build that
I don't think it's practical to introduce it as a generic package.
Hardware vendors continue to be bad at software.