coroutine: perform gc if no work is available
When polling for work using poll(2), first perform a non-blocking poll.
If no work is available, perform a major GC and then do a blocking poll.
This minimizes the heap size of the process when it is starved for work.

(It may not be optimal to try to perform a major GC each time this happens,
but we can add optimization heuristics later if it becomes necessary.)