plan: simplify sandbox dependencies
Don't depend on execline-tools being present
in or out of the sandbox; simply exec "/build"
inside the sandbox root after explicitly
performing the right fd manipulation.
Additionally, hard-code the environment
on entry to the sandbox. Build scripts can
modify it as necessary, so there's really
no need to make it adjustable.

Also, redirect the sandbox stdout and stderr
to zstd directly, rather than to an intermediate
file that gets compressed after-the-fact.
(This ought to reduce needless disk traffic.)