base: add exportall(1) (new package)
Port a tool for setting environment variables in bulk
inside execline scripts and port it into base.scm.

Add an 'env:' field to <package> and automatically
rewrite it to an exportall(1) invocation near the
head of the generated package build script.
Move some other sources of export(1) expressions
to exportall(1) expressions as well.

This changes build scripts across the world
of packages, but doesn't change any outputs.
(Verified with a rebuild of the x86_64 and aarch64 worlds.)