base: enable link-time dead code elimination
Compile with -ffunction-sections -fdata-sections -Wl,--gc-sections.

This reduces the *compressed* size of my simple VM rootfs
from 57 to 54MB, and it reduces the size of the prebuilt gcc
for x86_64 from ~31MB to ~30MB. So somewhere in the ballpark
of a 3-5% savings in binary sizes.

Confirmed that building the world on x86_64+aarch64+ppc64le works.