service: remount /proc in init
For reasons I can't discern, mounting /proc with hidepid=2
doesn't take the first time. (The mount command exits zero,
but the hidepid option doesn't show up in /proc/mounts, and
manual testing reveals that the feature is off.)

Remounting /proc with hidepid=2 appears to work.
(I'm able to repro this bug _and_ confirm that the
fix works in both kvm and on real hardware on two
different kernels...)