pkg/ip-wait: new package
Add a utility script for waiting on networking events.

This covers the simple cases where services need to
wait for networking to be 'up' (for some definition of up)
before starting in earnest. Searching for a pattern in
ip(8) output keeps us from having to make a global decision
about what it means for the network to be 'up,' and also
keeps the explicit service dependencies much simpler,
since we don't have to encode the notion of networking states
into the service DAG.

By default, wait for a default route to be installed
before starting ntpd.