all: change package representation
The <package> abstraction in package.scm was a little
odd, because it required a new kvector type to be
used as part of the intermediate representation even
though we threw it out almost immediately.

Instead, make "package" objects just functions
that expand (perhaps recursively) into artifacts or plans.
This makes it easier to add odd pseudo-packages, sub-packages, etc.

In order to make sure memoization works as intended,
put the 'build' config inside the 'host' config so that
we can't accidentally used a memoized plan/artifact when
the build config has changed. (Stuffing the build config
in a global parameter was a hack.)

Since this removes <package>, the bootstrap logic is
re-worked a bit. Now everything is explicitly built
as a two-stage bootstrap: the tools for building
the host config are built using the build config tools,
which are in turn built with prebuilt tools.
(This means the host tools ought to reproduce themselves.)

I've also checked in the (trivial) sub-package wrapper code,
but I haven't deployed it anywhere yet.