fs: swap parent/child relationship for kmsg and s6-log
The kmsg service was refusing to shut down cleanly on SIGTERM.
I gather this is because s6-log wants to see an EOF.

Swap the parent/child relationship in the kmsg run script
so that the SIGTERM is sent to the sed(1) process, which
will exit and subsequently cause s6-log to see an EOF
and exit cleanly.

Tested inside qemu.