plan.scm: create plan outputs symlink
Presently, there is no efficient way of determining
which plan produced an output, which means that
when we load a plan from disk, we are oblivious to
how its inputs were computed.

Create a symlink in the plan directory that lets
us address plans by output. (Keep the symlink pointed
at the most-recently-built version of a plan if there
are multiple versions that produce one output.)

This ought to let us reconstitute a complete plan graph
for a particular output without the code that generated
the original graph. (Note that the graph is approximate
when there are multiple plans that produce the same output.)