plat/droplet: use vda for /var
Droplets come with at least 25GB available on /dev/vda,
so it makes sense to use that for /var rather than any
additional attached volumes.

In order to produce /dev/vda3, we execute a script (/sbin/preboot)
very early in the boot process that reformats the disk
and reboots as necessary.

Also, improve the droplet jq networking script to actually
parse the netmask provided by the instance metadata, rather
than hard-coding it.

Also, fix a bug in ip-wait.