system, platform: rework
Major rework of the notion of a "platform."

Now running 'distill build <platform>' takes the
system definition in "system.scm" and builds it
for the given platform (expected to be found at
plat/<platform>.scm in the package search path).

The result of building for a platform is always a
bootable disk image. (I've manually tested every
platform under plat/.)

Also, the handling of /var was reworked so that
it is possible to extend the boot disk with an
additional partition that becomes the /var mount.
(This is handled at the <platform> level, so regular
system definitions need not worry about mounting
/var any more.)