execline: eliminate non-semantic parens
This was an annoying and invasive change with
lots of code-churn, but thankfully most of it
was deleting parens.

The fact that execline s-expressions contained
parens that were sometimes semantic
(wrapping execline blocks) and sometimes not
(delimiting commands) meant that it was challenging
to write code to correctly expand templated execline
expressions. (See the old package#vargs and package#cmd*
business and its strange handling of blocks.)

Instead, simplify the execline printer so that
lists always indicate blocks, and add some simple
sugar that splices function results into scripts.

(There's certainly some clean-up that needs to be
done here, but this is the minimal *working* diff.)