package: expand-package -> package-template
Now that the package code can expand lambdas
in most contexts within expand-package, get rid
of the explicit lambda that wraps every expand-package
invocation and stick it inside the newly-renamed
package-template.

The "cross:" keyword for package-template is a bit
of a hack to ensure that there is a way to evaluate
a lambda with the host as the argument, and then have
its results evaluated against the build config.