plan, kvector: address complation issue
For reasons I have not yet determined,
chicken's type analyzer miscompiled "conform?"
in kvector.scm because it erroneously assumed
that the previous
"(let (... (contract (make-vector ktlen #f))) ..."
meant that "contract" only ever contained #f
elements, when in fact we store new values into
the vector shortly afterwards.

Fixing the above revealed another issue in plan.scm,
which is that the constructor for <input> cannot
depend on the "plan?" predicate.