plan: use a string buffer in with-interned-output
The canonicalized output for plan-hash isn't large
enough to be worth buffering to a file, and the common
case is that the plan manifest has already been evaluated
in a previous invocation of the tool.

Also, expanded some comments around the codebase because
I was feeling friendly to my future self.