service: some fixups
The latest s6-svscan does not accept the -S argument,
so remove it from the invocation in /sbin/init.

Pass -regenerate-import-libraries to $(CHICKEN)
so that we don't confuse Make.

Re-format some ill-formatted s-expressions.

Ensure all of the core packages are available
from within the `distill` binary by importing
them inside a bare `(let () ...)`