sandbox: handle weird file descriptor numbering
Sometimes we get unlucky and the pipe file descriptors
for the Make jobserver turn out exactly backwards,
in which case we need to juggle the file descriptors
around a bit.

Write a separate routine for setting the jobserver
fds to 5 and 6, respectively, and handle the weird
edgecases where the pipe fds are 6, 5, or 6+5, etc.