plat/nitro-ami-x86_64: populate keys + hostname
We can populate /root/.ssh/authorized_keys and
hostname automatically from the ec2 metadata service.

Also, modify ip-wait so that it doesn't print the
result of the matching expressions from 'grep -E ...'

Finally, update plan-test.scm so that it will
run correctly regardless of whether or not
test-artifacts/ or test-plans/ have been created yet.
Similarly, make sure those directories are empty
before the test begins so that the test actually
has work to do if we've already run it once.