system: allow custom users, groups, perms
Allow platform and system specifications
to include users, groups, and rootfs file
permissions modifications.

The mkimage function for platform-mkimage
now must accept a third argument, which is the
set of filesystem permission overrides to apply.
When changing the ownership of a file, the
spec can use either a user/group name or
a uid/gid; user and group names are looked up
in the list of explicit users and groups
from the platform and system specs.

The users and groups specificed for platforms
and systems currently only support hardcoded
uids and gids.