all: improvements to permissions rewriting
Unfortunately mksquashfs ends up overriding
uids and gids present in the "pseudo-file"
specification when -all-root is passed.
The quickest solution for now is to patch
mksquashfs to fix the issue, but we should
see if maybe the behavior could be changed
upstream.

Also fix some buglets discovered in manual
testing.