all: add support for GC + user-fetch-hook
Up until now, the dependency resolution code
didn't draw a distinction between knowing the output
of a plan (i.e. its hash) and considering it "resolved"
from the standpoint of a build. Consequently, the code
would barf if the output of a plan was deleted but the
plan itself remained in the database.

So, update the logic to handle plans with known outputs
that are *not* present locally, and add user-fetch-hook
for trying to fetch plan outputs that aren't present.

Also, add the 'gc' subcommand for deleting all artifacts
that are not part of the dependency tree for a particular
build.

Also, simplify the coroutine lock (it was subtly buggy;
the new algorithm is fair and much simpler).