plan: handle flattened symlinks
If the 'output:hash' link isn't a symlink,
the current code will fail to handle it correctly.
(In the case where the plans directory is sync'd with
e.g. S3, the symlink will get flattened into a real directory.)

When removing an old 'output:hash' link, take care
to remove whatever happens to be there (directory, link, etc.)