all: simplify archive model
Rather than using bespoke code to implement
archival, just make archival part of the
build process. So, the plan.scm code itself
only deals with raw artifacts and not with
turning directories into artifacts.

The drawback here is that tar and zstd
have to be part of the prebuilt seed.
But, the advantages are significant:

 - No need to worry about new versions of
   zstd breaking reproducibility.
   The zstd version is captured explicitly
   because it's a 'tool' input.
 - No external dependencies for the distill
   binary itself.
 - Less code to maintain overall.

The subpackage constructor has also been
simplified a bit so that it doesn't require
any special features implemented in plan.scm;
it is just a plan that extracts a tarball and
then builds a smaller one from the extracted bits.