run: produce friendlier messages for eval exceptions
When `real-eval' produces an exception as we are loading
an import, annotate the error message with the import expression
that failed to load and produce an error via `print-error-message'.
Otherwise the load error is caught by the coroutine exception handler,
at which point valuable context is lost and we produce a significantly
worse error message.